+++ CHANGELOG.md
@@ -65,6 +65,7 @@
 - PIM-9974: [PIM-9972] Half the selected published product are unpublished through mass action
 - PIM-9984: Fix image preview in asset manager in media link attributes
 - PIM-9981: Fix permission to download logs of clean removed attribute values job
+- PIM-9980: Prevent root category selection in Rule Engine set categories action 
 
 ## Improvements